refactor: consolidate schema types into single module
- Moved common types (AccessToken, AnchoredTo, ContentBlock, etc.) from individual files into a unified schema structure
- Reorganized DataType module exports to reference new consolidated location
- Removed deprecated types and streamlined type organization for better maintainability
