+++ packages/sdk/schema/src/index.ts
@@ -3,11 +3,11 @@
 //
 
 export * from './annotations';
-export * from './common';
 export * from './construction';
 export * from './format';
 export * from './model';
 export * from './projection';
 export * from './properties';
+export * from './types';
 export * from './util';
 export * from './validate';
