+++ packages/sdk/schema/src/projection/projection.test.ts
@@ -26,7 +26,7 @@ import { EchoTestBuilder } from '@dxos/echo-db/testing';
 import { registerSignalsRuntime } from '@dxos/echo-signals';
 import { invariant } from '@dxos/invariant';
 
-import { DataType } from '../common';
+import { DataType } from '../types';
 
 import { createFieldId } from './field';
 import { ProjectionModel } from './projection';
