+++ packages/sdk/schema/src/types/AccessToken.ts
