+++ packages/sdk/schema/src/types/HasConnection.ts
