+++ packages/sdk/schema/src/types/HasRelationship.ts
