+++ packages/sdk/schema/src/types/Text.ts
