+++ packages/sdk/schema/src/types/util.ts
